tracing: just control ansi colors with feature flags
