+++ linkerd/app/inbound/fuzz/Cargo.toml
@@ -20,8 +20,6 @@ linkerd-app-inbound = { path = ".." }
 linkerd-app-test = { path = "../../test" }
 linkerd-tracing = { path = "../../../tracing", features = ["ansi"] }
 tracing = "0.1"
-tracing-subscriber = "0.2"
-
 # Prevent this from interfering with workspaces
 [workspace]
 members = ["."]
