+++ linkerd/app/inbound/src/http/tests.rs
@@ -14,6 +14,7 @@ use linkerd_app_core::{
     transport::{ClientAddr, Remote, ServerAddr},
     Conditional, NameAddr, ProxyRuntime,
 };
+use linkerd_tracing::test::trace_init;
 use tracing::Instrument;
 
 fn build_server<I>(
@@ -49,7 +50,7 @@ async fn unmeshed_http1_hello_world() {
     let mut server = hyper::server::conn::Http::new();
     server.http1_only(true);
     let mut client = ClientBuilder::new();
-    let _trace = support::trace_init();
+    let _trace = trace_init();
 
     let accept = HttpAccept {
         version: proxy::http::Version::Http1,
@@ -96,7 +97,7 @@ async fn downgrade_origin_form() {
     server.http1_only(true);
     let mut client = ClientBuilder::new();
     client.http2_only(true);
-    let _trace = support::trace_init();
+    let _trace = trace_init();
 
     let accept = HttpAccept {
         version: proxy::http::Version::H2,
@@ -144,7 +145,7 @@ async fn downgrade_absolute_form() {
     server.http1_only(true);
     let mut client = ClientBuilder::new();
     client.http2_only(true);
-    let _trace = support::trace_init();
+    let _trace = trace_init();
 
     let accept = HttpAccept {
         version: proxy::http::Version::H2,
