+++ linkerd/app/integration/Cargo.toml
@@ -28,6 +28,7 @@ linkerd-app-core = { path = "../core" }
 linkerd-metrics = { path = "../../metrics", features = ["test_util"] }
 linkerd2-proxy-api = { git = "https://github.com/linkerd/linkerd2-proxy-api", tag = "v0.1.18", features = ["arbitrary"] }
 linkerd-app-test = { path = "../test" }
+linkerd-tracing = { path = "../../tracing" }
 regex = "1"
 socket2 = "0.3.12"
 rustls = "0.19"
@@ -38,18 +39,8 @@ tower = { version = "0.4.7", default-features = false}
 tonic = { version = "0.4", default-features = false }
 tracing = "0.1.23"
 webpki = "0.21"
-
-[dependencies.tracing-subscriber]
-version = "0.2.14"
-# we don't need `chrono` time formatting or ANSI colored output
-default-features = false
-features = ["env-filter", "fmt", "smallvec", "tracing-log", "json", "parking_lot"]
+tracing-subscriber = { version = "0.2", default-features = false, features = ["fmt"]}
 
 [dev-dependencies]
 flate2 = { version = "1.0.1", default-features = false, features = ["rust_backend"] }
-
-[dev-dependencies.tracing-subscriber]
-version = "0.2.14"
-# turn on ANSI colors for tests :)
-default-features = false
-features = ["env-filter", "fmt", "smallvec", "ansi", "tracing-log", "json", "parking_lot"]
+linkerd-tracing = { path = "../../tracing", features = ["ansi"] }
