+++ linkerd/app/integration/src/lib.rs
@@ -24,6 +24,7 @@ pub use linkerd_app::{
     core::{drain, Addr},
 };
 pub use linkerd_app_test::*;
+pub use linkerd_tracing::test::*;
 use socket2::Socket;
 pub use std::collections::HashMap;
 use std::fmt;
@@ -40,7 +41,6 @@ pub use tokio::sync::oneshot;
 pub use tonic as grpc;
 pub use tower::Service;
 pub use tracing::*;
-pub use tracing_subscriber::prelude::*;
 
 /// Environment variable for overriding the test patience.
 pub const ENV_TEST_PATIENCE_MS: &str = "RUST_TEST_PATIENCE_MS";
