+++ linkerd/app/outbound/src/tcp/opaque_transport.rs
@@ -165,8 +165,7 @@ mod test {
 
     #[tokio::test(flavor = "current_thread")]
     async fn plain() {
-        #[cfg(feature = "test-subscriber")]
-        let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+        let _trace = linkerd_tracing::test::trace_init();
 
         let svc = OpaqueTransport {
             inner: service_fn(|ep: Connect| {
@@ -188,8 +187,7 @@ mod test {
 
     #[tokio::test(flavor = "current_thread")]
     async fn opaque_no_name() {
-        #[cfg(feature = "test-subscriber")]
-        let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+        let _trace = linkerd_tracing::test::trace_init();
 
         let svc = OpaqueTransport {
             inner: service_fn(|ep: Connect| {
@@ -227,8 +225,7 @@ mod test {
 
     #[tokio::test(flavor = "current_thread")]
     async fn opaque_named_with_port() {
-        #[cfg(feature = "test-subscriber")]
-        let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+        let _trace = linkerd_tracing::test::trace_init();
 
         let svc = OpaqueTransport {
             inner: service_fn(|ep: Connect| {
@@ -266,8 +263,7 @@ mod test {
 
     #[tokio::test(flavor = "current_thread")]
     async fn opaque_named_no_port() {
-        #[cfg(feature = "test-subscriber")]
-        let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+        let _trace = linkerd_tracing::test::trace_init();
 
         let svc = OpaqueTransport {
             inner: service_fn(|ep: Connect| {
