+++ linkerd/proxy/http/Cargo.toml
@@ -44,4 +44,4 @@ tokio-test = "0.4"
 
 [dev-dependencies]
 tokio-test = "0.4"
-tracing-subscriber = "0.2.16"
+linkerd-tracing = { path = "../../tracing", features = ["ansi"] }
