+++ linkerd/tls/fuzz/fuzz_targets/fuzz_target_1.rs
@@ -5,6 +5,8 @@ use libfuzzer_sys::fuzz_target;
 
 #[cfg(fuzzing)]
 fuzz_target!(|data: &[u8]| {
-    let _trace = linkerd_tracing::test::trace_init();
+    // Don't enable tracing in `cluster-fuzz`, since we would emit verbose
+    // traces for *every* generated fuzz input...
+    let _trace = linkerd_tracing::test::with_default_filter("off");
     linkerd_tls::server::fuzz_logic::fuzz_entry(data);
 });
