+++ linkerd/tracing/Cargo.toml
@@ -6,6 +6,10 @@ license = "Apache-2.0"
 edition = "2018"
 publish = false
 
+[features]
+default = []
+ansi = ["tracing-subscriber/ansi"]
+
 [dependencies]
 bytes = "1"
 html-escape = "0.2.5"
@@ -21,5 +25,5 @@ tracing-log = "0.1.2"
 [dependencies.tracing-subscriber]
 version = "0.2.16"
 default-features = false
-features = ["ansi", "env-filter", "fmt", "smallvec", "tracing-log", "json", "parking_lot"]
+features = ["env-filter", "fmt", "smallvec", "tracing-log", "json", "parking_lot"]
